+++ src/configuration/mod.rs
@@ -1,3 +1,3 @@
-pub mod configuration;
+pub mod app_configuration;
 pub mod keybindings;
 pub mod styles;
