+++ src/tui/mod.rs
@@ -1,4 +1,4 @@
 pub mod action;
 pub mod components;
 pub mod mode;
-pub mod tui;
+pub mod ui;
